Move sets to apis/run/util and compatibility to version. Skip failing test
