+++ apis/run/util/version/compatibility.go
@@ -1,12 +1,14 @@
 // Copyright 2022 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package sets
+package version
 
 import (
 	"context"
+
+	"github.com/vmware-tanzu/tanzu-framework/apis/run/util/sets"
 )
 
 type Compatibility interface {
-	CompatibleVersions(ctx context.Context) (StringSet, error)
+	CompatibleVersions(ctx context.Context) (sets.StringSet, error)
 }
